test: enhance alias handling tests and refactor test helpers for clarity
